refactor: refactor e2e tests
