+++ cypress/integration/02-approvevenft.test.ts
@@ -2,28 +2,14 @@ import {
   HasVeNFTToSellApprovedAllBridge,
   HasVeNFTToSellBridge,
   provider,
-  SellVeNFTBridge,
   signer,
   ZeroBalanceVeNFTBridge,
 } from '../support/commands'
-import { TEST_ADDRESS_NEVER_USE_SHORTENED, tokenListSorted } from '../utils/data'
+import { tokenListSorted } from '../utils/data'
 import { Vault } from '../../src/constants/addresses'
 import { SupportedChainId } from '../../src/constants/chains'
 
-describe('VeNFT Sell', () => {
-  const setupEthBridge = () => {
-    cy.on('window:before:load', (win) => {
-      // @ts-ignore
-      win.ethereum = new ZeroBalanceVeNFTBridge(signer, provider)
-    })
-  }
-
-  it('is connected', () => {
-    setupEthBridge()
-    cy.visit('/venft/sell/')
-    cy.get('[data-testid=wallet-connect]', { timeout: 1000 }).contains(TEST_ADDRESS_NEVER_USE_SHORTENED)
-  })
-
+describe('Approve VeNFT', () => {
   it('gets VeNFT balance', () => {
     const ethBridge = new ZeroBalanceVeNFTBridge(signer, provider)
     cy.on('window:before:load', (win) => {
@@ -52,7 +38,7 @@ describe('VeNFT Sell', () => {
 
     cy.visit('/venft/sell/')
     const expectTokenId = tokenListSorted[1].tokenId
-    cy.get(`[data-testid=venft-sell-row-1-token-id]`, { timeout: 2000 }).contains(expectTokenId)
+    cy.get(`[data-testid=venft-sell-row-1-token-id]`).contains(expectTokenId)
     cy.window().then((win) => {
       expect(ethBridge.VeNFTBalanceOf).to.have.callCount(1)
       expect(ethBridge.tokenOfOwnerByIndex).to.have.callCount(tokenListSorted.length)
@@ -96,25 +82,6 @@ describe('VeNFT Sell', () => {
     cy.wait(1000)
     cy.window().then((win) => {
       expect(ethBridge.approveSpy).to.have.calledWith(Vault[SupportedChainId.FANTOM], tokenListSorted[2].tokenId)
-    })
-    cy.get('[data-testid=explorer-link-success-box]')
-  })
-
-  it('sells VeNFT', () => {
-    const ethBridge = new SellVeNFTBridge(signer, provider)
-    cy.on('window:before:load', (win) => {
-      // @ts-ignore
-      win.ethereum = ethBridge
-      cy.spy(ethBridge, 'sellVeNFTSpy')
-    })
-
-    cy.visit('/venft/sell/')
-    const expectTokenId = tokenListSorted[1].tokenId
-    cy.get(`[data-testid=venft-sell-row-1-action]`).contains('Sell')
-    cy.get(`[data-testid=venft-sell-row-1-action]`).click()
-    cy.wait(500)
-    cy.window().then((win) => {
-      expect(ethBridge.sellVeNFTSpy).to.have.calledWith(expectTokenId)
     })
     cy.get('[data-testid=explorer-link-success-box]')
   })
